Resolve MD032 markdown lint errors
'MD032 Lists should be surrounded by blank lines', can occur when a newline separator is missing after markdown lists.

For example,

* zomg
* apples

Will fail due to no newline between the list and the heading 'Some other content'.

This is purely stylistic, but makes editing these files nicer.

* zomg
* apples

Will not fail.

Additionally, this may occur when line wrapping occurs between list items without proper indentation.

For example,

* zomg
this is the same list item as above
* but the parser thinks this is a new list

Will fail, unless we indent the second line.

* zomg
  this is the same list item..
* but this works

Will not fail, since the list item is indented properly.

Alternatively, we can unwrap the line to a single-long form line to pass this lint rule.

This is probably preferred, since wrapping is purely superficial and is not entirely necessary.

The benefit to this approach is that the writer can choose if they want to line-wrap, they can, just as long as the subsequent lines are indented within the same list item. Otherwise, just unwrap.